requirements.api.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File handling
aiofiles==23.2.1
PyPDF2==3.0.1
python-docx==1.1.0

# Environment variables
python-dotenv==1.0.0

# Data validation
pydantic==2.5.0

# HTTP client
httpx==0.25.2

# Logging
structlog==23.2.0

# Database (optional)
sqlalchemy==2.0.23
alembic==1.13.1

# CORS middleware
fastapi-cors==0.0.6

# Optional LLM providers
# openai>=1.0.0
# google-generativeai>=0.3.0

# Additional utilities
python-dateutil==2.8.2
typing-extensions==4.8.0